pyproject.toml
[project]
name = "gtfs-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "dotenv>=0.9.9",
    "fastapi",
    "fastapi[standard]",
    "google-adk",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "sqlacodegen>=4.0.3",
    "sqlalchemy>=2.0.48",
    "uvicorn",
]
